refactor: convert automation/server.py into a mixin package (#536)
Split the ~1112-LOC AutomationServer monolith into an automation/server/
package mirroring controllers/app_controller/, using the established
TYPE_CHECKING _Base = Proto else object mixin composition style.

- server/__init__.py composes AutomationServer from the mixins and keeps the
  public surface (AutomationServer, DEFAULT_PORT, BUFFER_SIZE) on the same
  automation.server import path. The handler-registry dict is preserved
  byte-for-byte (36 commands, identical names → handler methods).
- server/window_capture.py: stateless Win32/ctypes setup + Pillow PNG save
  helper (non-mixin support module).
- server/protocol.py: AutomationServerProto typing the composed host.
- Mixins: TransportMixin, ScreenshotMixin, IntrospectionMixin,
  DeckResearchMixin, ZoneEditingMixin, BuilderMixin, ManaRenderingMixin.

Behavior-preserving: no change to the wire protocol, command names, port,
or dispatch behavior.
